Add soft-delete support for milestones
Introduce soft-delete operations and related fixes for milestone deletion flow. Key changes:
- Add IMilestoneRepository.GetByIdWithDeletedAsync and SoftDeleteAsync to the repository interface.
- Implement GetByIdWithDeletedAsync (IgnoreQueryFilters + Include) and SoftDeleteAsync (ExecuteUpdateAsync setting IsDeleted, DeletedAt, DeletedBy) in MilestoneRepository.
- Change Milestone.DeletedBy type from int to long to store user chat IDs.
- Update ConfirmDeleteStepHandler to call SoftDeleteAsync, handle a 0-result failure (notify user, refresh list, reset state) and return Idle on failure; keep existing success flow returning DeleteMilestoneWaitingUndo.
- Update UndoDeleteStepHandler to use GetByIdWithDeletedAsync when restoring.

These changes centralize and optimize deletion updates in the database, ensure deleted records can be queried when needed, and properly handle failures in the UI flow.

+++ MilestoneTracker.Domain/Entities/Milestones/Milestone.cs
@@ -13,7 +13,7 @@ public class Milestone
     public string Title { get; set; } = null!;
     public string? Description { get; set; }
     public bool IsDeleted { get; set; } = false;
-    public int? DeletedBy { get; set; }
+    public long? DeletedBy { get; set; }
     public DateTime? DeletedAt { get; set; }
